news, careers, events fields changed to textfield from checkbox
Idealy it shoud be checkbox. But when I create a view for checkbox fields, it has two value either true or empty. But when I add checkbox field for filtering, it gives any, ture and false option. the false should corresponds empty but it does not. So, I have can not filter in or out empty fields. So, I am sending Yes for true and No for false value. So that I can filter in or out from wiev based on Yes or No value.